Refactor and modernize library and add linting
