+++ phpunit.xml
@@ -1,17 +1,26 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<phpunit backupGlobals="true"
-         backupStaticAttributes="true" 
-         bootstrap="vendor/autoload.php" 
-         colors="true"
-         convertErrorsToExceptions="true" 
-         convertNoticesToExceptions="true" 
-         convertWarningsToExceptions="true" 
-         processIsolation="false" 
-         stopOnFailure="false" 
-         syntaxCheck="true">
+<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/9.5/phpunit.xsd"
+         bootstrap="vendor/autoload.php"
+         cacheResultFile=".phpunit.cache/test-results"
+         executionOrder="depends,defects"
+         forceCoversAnnotation="false"
+         beStrictAboutCoversAnnotation="true"
+         beStrictAboutOutputDuringTests="true"
+         beStrictAboutTodoAnnotatedTests="true"
+         failOnRisky="true"
+         failOnWarning="false"
+         verbose="true">
     <testsuites>
-    	<testsuite name="PHPHelpers Test Suite">
-    		<directory suffix=".php">tests</directory>
-    	</testsuite>
+        <testsuite name="default">
+            <directory suffix="Test.php">tests</directory>
+        </testsuite>
     </testsuites>
+
+    <coverage cacheDirectory=".phpunit.cache/code-coverage"
+              processUncoveredFiles="true">
+        <include>
+            <directory suffix=".php">src</directory>
+        </include>
+    </coverage>
 </phpunit>
